Add files via upload
The last PR was wrong, this one is right.
I have modified my code which refers to the "fix ave/time" command in LAMMPS. Specifically, I have configured three parameters: Nevery, Nrepeat, and Nfreq. These parameters allow me to select a subset of frames for calculation. For instance, with Nevery set to 2, Nrepeat set to 6, and Nfreq set to 100, the code will choose frames at intervals of 90, 92, 94, 96, 98, 100, and subsequently at 190, 192, 194, 196, 198, 200, and so on. 
I have changed the way to read the frame. I only read the coordinates which are going to be calculated